Chris/development (#32)
* week 3

* Create Bounded_waves.ipynb

* Add files via upload

* Add files via upload

* Add files via upload

* Update Week_3_Initialize.ipynb

* Update Week_3_Initialize.ipynb

* Update Week_3_Initialize.ipynb

* Create Ipython_error_in_Git_Lab.ipynb

* trim nbs

* cookbook update weeks

* updates

* Update cookbook.ipynb

* Update cookbook.ipynb

* unpack list of widgets

* update cookbook

included basic functions, that can be used to learn and build widgets. A summary / cleaned up version will follow.

* Reworked cookbook

The cookbook now consists of 3 parts:
- An overview of the usable question types
- The realization of the question types, with code explanation
- The most important results of the preliminary study (pre-summer)

* Delete notebooks/Chris/Initialize/Week_3_Initialize.ipynb

* Delete notebooks/Chris/Initialize/Week_5_Initialize.ipynb

* Delete notebooks/Chris/Ipython_error_in_Git_Lab.ipynb

* Delete notebooks/Chris/Week 3.ipynb

* Delete notebooks/Chris/week 5.ipynb

* Start week2

* Improvement button widgets

* Fix/restore (submit) button

* Various Updates

week 2, 5, and cookbook

* Include coding questions in cookbook, both parameters and functions

New question types are included in the cookbook. Ask students to code questions and let them check if the calculated values are correct or plot their function to the correct function. Here it checks if the function is valid, and if so if the outcome is good or not.

* Backup working functions

This is the final version of a graph, which is not yet used. First more basic figures will be introduced to students.

* reupload cookbook, begin question coding parameter week 2

* parameter coding question and week 2

In the coding question, where students have to calculate one or more parameters, it is now allowed to have a small computational error of 0.01%, on default, which can be caused by iterative approaches. This is applied in week 2.

* Improvement coding questions in cookbook and week2

Functions for coding questions have been more generalized, easing the making of similar questions. These are applied in week 2.

---------

Co-authored-by: Christiaan Hoogervorst <C.D.Hoogervorst@tudelft.nl>
Co-authored-by: Floris Calkoen <44444001+FlorisCalkoen@users.noreply.github.com>
Co-authored-by: floriscalkoen <floris_calkoen@hotmail.com>