Feedback progress - shared initialize file
Feedback from 2a, 2b (up to question 7) is processed. A (personal) review is still needed.

A shared notebook is added. These functions should get into the coastal package. Changes are made to the functions, and precede the code in the cookbook. (No changes are required for code that calls  these functions)