Update 2c_theory_equilibrium_tide.ipynb
Hi Mia/Judith:

I just looked at 2c and fixed a few typos, added a few cross-references to the textbook, and put variable names like g in latex. I see that there are still some (unaddressed?) notes from Judith (code section 4) about whether or not we give the students those values.

I also added a little bit of text to the bottom to wrap things up (it seemed to end a little abruptly):
“By adjusting the above figure, we can see how the alignment of the earth, moon, and sun influences the tide experienced by a given point on earth.  If the tidal bulges align, we get a higher high tide, "spring tide", but if the tidal bulges are out of alignment, we have a reduced high tide, the "neap tide". The same holds for low waters: low tides are lower when everything is aligned (spring tide), and the low tides at neap are not quite as low.”

All in all it looked great, and I thought this really accomplished our goal of bringing to life some of the trickier concepts in the textbook.

Cheers,

Stuart